api_apps/retrieve-tweets/requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
idna==2.10
joblib==1.0.0
nltk==3.5
numpy==1.19.4
oauthlib==3.1.0
pandas==1.1.5
pip==20.3.3
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
setuptools==49.2.1
six==1.15.0
tqdm==4.54.1
twython==3.8.2
urllib3==1.26.2
